:recycle: (pc-13307) (bsr) move string checkers in fraud utils
